Conduct Peak Annotation.

Conduct Peak Annotation.

AnnoPeak(
  peak.df,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  seq.style = c("UCSC", "NCBI", "Ensembl", "None"),
  gtf.file = NULL,
  up.dist = 3000,
  down.dist = 3000,
  ...
)

Arguments

peak.df
Dataframe contains all consensus peaks.

species
Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis","Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus", "Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".

seq.style
The style of sequence, chosen from UCSC, NCBI, Ensembl, None. This should be compatible with the genome and gtf file you used to generate count matrix and peak files. Default: "UCSC".

gtf.file
GTF file used to create TxDb object. Useful when specie you used is not available in species. Default: NULL.

up.dist
The upstream distance from the TSS. Default: 3000bp.

down.dist
The downstream distance from the TSS. Default: 3000bp.

...
Parameters for annotatePeak.

Value

List contains peak annotation dataframe and plot.

Examples

library(DEbPeak)
peak.file <- system.file("extdata", "debchip_peaks.bed", package = "DEbPeak")
peak.df <- GetConsensusPeak(peak.file = peak.file)
peak.profile <- PeakProfile(peak.df, species = "Mouse", by = "gene", region.type = "body", nbin = 800)

#> 
#> 载入程辑包：‘BiocGenerics’

#> The following objects are masked from ‘package:stats’:
#> 
#>     IQR, mad, sd, var, xtabs

#> The following objects are masked from ‘package:base’:
#> 
#>     anyDuplicated, append, as.data.frame, basename, cbind, colnames,
#>     dirname, do.call, duplicated, eval, evalq, Filter, Find, get, grep,
#>     grepl, intersect, is.unsorted, lapply, Map, mapply, match, mget,
#>     order, paste, pmax, pmax.int, pmin, pmin.int, Position, rank,
#>     rbind, Reduce, rownames, sapply, setdiff, sort, table, tapply,
#>     union, unique, unsplit, which.max, which.min

#> 
#> 载入程辑包：‘S4Vectors’

#> The following object is masked from ‘package:base’:
#> 
#>     expand.grid

#> Welcome to Bioconductor
#> 
#>     Vignettes contain introductory material; view with
#>     'browseVignettes()'. To cite Bioconductor, see
#>     'citation("Biobase")', and for packages 'citation("pkgname")'.

#> >> preparing promoter regions...	 2022-12-12 11时57分35秒 
#> >> preparing tag matrix...		 2022-12-12 11时57分36秒 
#> >> preparing start_site regions by ... 2022-12-12 11时57分36秒
#> >> preparing tag matrix...  2022-12-12 11时57分36秒 
#> >> generating figure...		 2022-12-12 11时57分46秒 

#> >> done...			 2022-12-12 11时57分46秒 

#> >> binning method is used...2022-12-12 11时57分46秒
#> >> preparing start_site regions by gene... 2022-12-12 11时57分46秒
#> >> preparing tag matrix by binning...  2022-12-12 11时57分46秒 
#> >> Running bootstrapping for tag matrix...		 2022-12-12 11时57分50秒 
#> >> binning method is used...2022-12-12 11时57分50秒
#> >> preparing body regions by gene... 2022-12-12 11时57分50秒
#> >> preparing tag matrix by binning...  2022-12-12 11时57分50秒 
#> >> preparing matrix with extension from (TSS-20%)~(TTS+20%)... 2022-12-12 11时57分50秒
#> >> 1 peaks(0.1536098%), having lengths smaller than 800bp, are filtered... 2022-12-12 11时57分53秒
#> >> Running bootstrapping for tag matrix...		 2022-12-12 11时58分30秒 

peak.anno <- AnnoPeak(
  peak.df = peak.df, species = "Mouse",
  seq.style = "UCSC", up.dist = 20000, down.dist = 20000
)

#> 

#> >> preparing features information...		 2022-12-12 11时58分31秒 
#> >> identifying nearest features...		 2022-12-12 11时58分32秒 
#> >> calculating distance from peak to TSS...	 2022-12-12 11时58分32秒 
#> >> assigning genomic annotation...		 2022-12-12 11时58分32秒 
#> >> adding gene annotation...			 2022-12-12 11时58分44秒 

#> 'select()' returned 1:many mapping between keys and columns

#> >> assigning chromosome lengths			 2022-12-12 11时58分44秒 
#> >> done...					 2022-12-12 11时58分44秒 

#> Warning: Removed 6 rows containing non-finite values (stat_count).