Create Quadrant Diagram for Differential Expression Analysis of RNA-seq and Peak-related Data.

Create Quadrant Diagram for Differential Expression Analysis of RNA-seq and Peak-related Data.

InteDiffQuad(
  de.peak,
  peak.type = c("ChIP", "ATAC"),
  point.alpha = 0.6,
  point.size.vec = c(2, 4),
  rna.l2fc.threshold = 0,
  peak.l2fc.threshold = 0,
  linetype = 2,
  point.color.vec = c("grey", "red", "blue", "grey"),
  legend.pos = "top",
  show.corr = TRUE,
  label.num = NULL,
  label.df = NULL,
  label.color = NULL
)

Arguments

de.peak
Dataframe contains integrated results of differential analysis of RNA-seq and Peak-related data.

peak.type
Peak data type, choose from ChIP, ATAC. Default: ChIP.

point.alpha
Opacity of a geom. Default: 0.6.

point.size.vec
Point size for regular(DE or non-DE) and(or) labeled points. Default: 2 for regular and 4 for labeled points.

rna.l2fc.threshold
Log2 fold change threshold for RNA-seq to get differentially expressed results. Default: 0.

peak.l2fc.threshold
Log2 fold change threshold for peak-related data to get differentially accessible/binding peaks. Default: 0.

linetype
Threshold linetype. Default: 2.

point.color.vec
Point color for Down_Up, Up_Up, Down_Down, Up_Down. Default: grey for Down_Up, red for Up_Up, blue for Down_Down and grey for Up_Down.

legend.pos
Legend position. Default: top.

show.corr
Logical value, whether to add pearson correlation coefficient and its significance. Default: TRUE.

label.num
Gene number to label, according to RNA_log2FoldChange and Peak_log2FoldChange. When label.df is set NULL, use this to determine genes to label. Default: NULL.

label.df
Label data frame, at least contains Gene column. Default: NULL(use label.num). When provided, the second column should not be in de.peak.

label.color
Color for labels. Default: NULL (black).

Value

A ggplot2 object.

Examples

# # prepare integrated differential expression results of RNA-seq and ATAC-seq
# # label dataframe
# label.df = data.frame(Gene = c("Ccl3", "Ccl5", "Cd28", "Cx3cr1", "Prdm1", "Tcf7", "Slamf6", "Id3", "Cxcr5"))
# # create plot
# quad.plot = InteDiffQuad(de.peak = debatac.res, peak.type = "ATAC", label.df = label.df, label.color = "green")