Get Near Genes of Differential Peaks.

Get Near Genes of Differential Peaks.

ProcessEnhancer(
  de.res,
  signif = "padj",
  signif.threshold = 0.05,
  l2fc.threshold = 1,
  label.key = NULL,
  species = c("Human", "Mouse", "Rat", "Fly", "Arabidopsis", "Yeast", "Zebrafish",
    "Worm", "Bovine", "Pig", "Chicken", "Rhesus", "Canine", "Xenopus", "Anopheles",
    "Chimp", "E coli strain Sakai", "Myxococcus xanthus DK 1622"),
  seq.style = "Ensembl",
  gtf.file = NULL,
  dis.threshold = 250000,
  n.cores = 1
)

Arguments

de.res
Dataframe contains all differential analysis results of peak-related data.

signif
Significance criterion for RNA-seq results. For DESeq2 results, can be chosen from padj, pvalue. For edgeR results, can be chosen from FDR, PValue. Default: padj.

signif.threshold
Significance threshold for RNA-seq to get differentially expressed genes. Default: 0.05.

l2fc.threshold
Log2 fold change threshold for RNA-seq to get differentially expressed genes. Default: 1.

label.key
Which column to use as label. Default: NULL (use rownames of de.res).

species
Species used, chosen from "Human","Mouse","Rat","Fly","Arabidopsis", "Yeast","Zebrafish","Worm","Bovine","Pig","Chicken","Rhesus", "Canine","Xenopus","Anopheles","Chimp","E coli strain Sakai","Myxococcus xanthus DK 1622". Default: "Human".

seq.style
The style of sequence, chosen from UCSC, NCBI, Ensembl, None. This should be compatible with the genome and gtf file you used to generate count matrix and peak files. Default: "UCSC".

gtf.file
GTF file used to create TxDb object. Useful when specie you used is not available in species. Default: NULL.

dis.threshold
Distance threshold. Default: 250000.

n.cores
The number of cores to be used for this job. Default:1.

Value

Dataframe contains differential peaks and their near genes.

Examples

# # multi-core (faster)
# enchancer.genes = ProcessEnhancer(de.res = all.enhancer.res, signif = "pvalue", signif.threshold = 0.05,
#                                   l2fc.threshold = 0, label.key = NULL, species = "Mouse", n.cores = 5)
# # single-core (slower)
# enchancer.genes2 = ProcessEnhancer(de.res = all.enhancer.res, signif = "pvalue", signif.threshold = 0.05,
#                                    l2fc.threshold = 0, label.key = NULL, species = "Mouse", n.cores = 1)