Identify TFs from Gene Expression Reults with ChEA3.

Identify TFs from Gene Expression Reults with ChEA3.

RunChEA3(
  genes,
  species = c("Human", "Mouse"),
  projet.name = NULL,
  library = c("meanRank", "topRank", "GTEx", "ReMap", "Enrichr", "ENCODE", "ARCHS4",
    "Literature")
)

Arguments

genes
Input gene symbol of human or mouse, can be a set of differentially expressed genes.

species
The species of input gene set. Choose from "Human" and "Mouse". Default: "Human".

projet.name
The name of the project used as Query Name. Default: NULL (DEbPeak).

library
The output library type, choose from "meanRank", "topRank", "GTEx", "ReMap", "Enrichr", "ENCODE", "ARCHS4", "Literature". Default: "meanRank".

Value

Dataframe contains identified TFs.

Examples

# library(DEbPeak)
# RunChEA3(genes = c("SMAD9","FOXO1","MYC","STAT1",'STAT3',"SMAD3"), species = "Human")