Create Quadrant Diagram for Two Differential Analysis Integration Results.

Create Quadrant Diagram for Two Differential Analysis Integration Results.

InteDiffQuad(
  inte.res,
  inte.type = c("DEbPeak", "PeakbPeak", "DEbDE"),
  point.alpha = 0.6,
  point.size.vec = c(2, 4),
  f1.l2fc.threshold = 0,
  f2.l2fc.threshold = 0,
  linetype = 2,
  point.color.vec = c("grey", "red", "blue", "grey"),
  legend.pos = "top",
  show.corr = TRUE,
  label.num = NULL,
  label.df = NULL,
  label.color = NULL
)

Arguments

inte.res
Integration results, can be output of DEbPeak, PeakbPeak, DEbDE.

inte.type
The integration type, choose from "DEbDE", "PeakbPeak", "DEbPeak". Default: "DEbPeak".

point.alpha
Opacity of a geom. Default: 0.6.

point.size.vec
Point size for regular and(or) labeled points (specified by label.df or label.num). Default: 2 for regular and 4 for labeled points.

f1.l2fc.threshold
Log2 fold change threshold for file1 to get differential analysis results. Default: 0.

f2.l2fc.threshold
Log2 fold change threshold for file2 to get differential analysis results. Default: 0.

linetype
Threshold linetype. Default: 2.

point.color.vec
Point color for Down_Up, Up_Up, Down_Down, Up_Down. Default: grey for Down_Up, red for Up_Up, blue for Down_Down and grey for Up_Down.

legend.pos
Legend position. Default: top.

show.corr
Logical value, whether to add pearson correlation coefficient and its significance. Default: TRUE.

label.num
Gene number to label, according to RNA_log2FoldChange and Peak_log2FoldChange. When label.df is set NULL, use this to determine genes to label. Default: NULL.

label.df
Label data frame, at least contains Gene column. Default: NULL(use label.num). When provided, the second column should not be in de.peak.

label.color
Color for labels. Default: NULL (black).

Value

A ggplot2 object.

Examples

# RNA-seq and RNA-seq
# de.de.label.df = data.frame(Gene = c("ENSDARG00000007396", "ENSDARG00000010729", "ENSDARG00000002194", "ENSDARG00000002587"))
# de.de.quad = InteDiffQuad(inte.res = de.de, inte.type = "DEbDE", f1.l2fc.threshold = 0.6, f2.l2fc.threshold = 0.6,
#                           show.corr = FALSE, label.df = de.de.label.df)
# Peak and Peak
# peak.peak.label.df = data.frame(Gene = c("Aak1", "Adam19", "Oaf", "Nsg2"))
# peak.peak.quad = InteDiffQuad(inte.res = atac.atac, inte.type = "PeakbPeak", f1.l2fc.threshold = 0,
#                               f2.l2fc.threshold = 0.5, show.corr = TRUE, label.df = peak.peak.label.df)
# RNA-seq and Peak
# de.peak.label.df = data.frame(Gene = c("Ccl3", "Ccl5", "Cd28", "Cx3cr1", "Prdm1", "Tcf7", "Slamf6", "Id3", "Cxcr5"))
# de.peak.quad = InteDiffQuad(inte.res = debatac.res, inte.type = "DEbPeak", f1.l2fc.threshold = 0,
#                             f2.l2fc.threshold = 0, show.corr = TRUE, label.df = de.peak.label.df)